Rename infer_bounds to slide_and_fold_storage
