+++ builder/BUILD
@@ -9,7 +9,6 @@ cc_library(
     name = "builder",
     srcs = [
         "pipeline.cc",
-        "infer_bounds.cc",
         "node_mutator.cc",
         "optimizations.cc",
         "rewrite.cc",
@@ -17,14 +16,15 @@ cc_library(
         "simplify.cc",
         "simplify_bounds.cc",
         "simplify_exprs.cc",
+        "slide_and_fold_storage.cc",
         "substitute.cc",
     ],
     hdrs = [
         "pipeline.h",
-        "infer_bounds.h",
         "node_mutator.h",
         "optimizations.h",
         "simplify.h",
+        "slide_and_fold_storage.h",
         "substitute.h",
     ],
     deps = ["//runtime"],
